+++ tests/test_async.py
@@ -17,7 +17,7 @@ from .utils import *
 
 
 pytest_plugins = ('pytest_asyncio',)
-pytestmark = pytest.mark.asyncio    # mark all tests
+pytestmark = pytest.mark.asyncio
 
 
 async def test_shield():