+++ tests/test_inputexp.py
@@ -15,7 +15,7 @@ import edzed
 from .utils import *
 
 pytest_plugins = ('pytest_asyncio',)
-pytestmark = pytest.mark.asyncio    # mark all tests
+pytestmark = pytest.mark.asyncio
 
 
 async def test_expiration(circuit):